+++ src/util/packed/mutable_enum.rs
@@ -44,7 +44,7 @@ impl Accountable for MutableEnum {
     }
 }
 impl Reader for MutableEnum {
-    fn get(&mut self, index: usize) -> Result<i64, LuceneError> {
+    fn get(&mut self, index: i32) -> Result<i64, LuceneError> {
         match self {
             MutableEnum::Packed(op) => op.get(index),
             MutableEnum::GrowableW(op) => op.get(index),
@@ -54,11 +54,11 @@ impl Reader for MutableEnum {
 
     fn get_bulk(
         &mut self,
-        index: usize,
+        index: i32,
         arr: &mut [i64],
-        off: usize,
-        len: usize,
-    ) -> Result<u32, LuceneError> {
+        off: i32,
+        len: i32,
+    ) -> Result<i32, LuceneError> {
         match self {
             MutableEnum::Packed(op) => op.get_bulk(index, arr, off, len),
             MutableEnum::GrowableW(op) => op.get_bulk(index, arr, off, len),
@@ -66,7 +66,7 @@ impl Reader for MutableEnum {
         }
     }
 
-    fn size(&self) -> u32 {
+    fn size(&self) -> i32 {
         match self {
             MutableEnum::Packed(op) => op.size(),
             MutableEnum::GrowableW(op) => op.size(),
@@ -75,7 +75,7 @@ impl Reader for MutableEnum {
     }
 }
 impl Mutable for MutableEnum {
-    fn get_bits_per_value(&self) -> u32 {
+    fn get_bits_per_value(&self) -> i32 {
         match self {
             MutableEnum::Packed(op) => op.get_bits_per_value(),
             MutableEnum::GrowableW(op) => op.get_bits_per_value(),
@@ -83,7 +83,7 @@ impl Mutable for MutableEnum {
         }
     }
 
-    fn set(&mut self, index: usize, value: i64) -> Result<(), LuceneError> {
+    fn set(&mut self, index: i32, value: i64) -> Result<(), LuceneError> {
         match self {
             MutableEnum::Packed(op) => op.set(index, value),
             MutableEnum::GrowableW(op) => op.set(index, value),
@@ -93,11 +93,11 @@ impl Mutable for MutableEnum {
 
     fn set_bulk(
         &mut self,
-        index: usize,
+        index: i32,
         arr: &[i64],
-        off: usize,
-        len: usize,
-    ) -> Result<u32, LuceneError> {
+        off: i32,
+        len: i32,
+    ) -> Result<i32, LuceneError> {
         match self {
             MutableEnum::Packed(op) => op.set_bulk(index, arr, off, len),
             MutableEnum::GrowableW(op) => op.set_bulk(index, arr, off, len),
@@ -105,7 +105,7 @@ impl Mutable for MutableEnum {
         }
     }
 
-    fn fill(&mut self, from_index: usize, to_index: usize, val: i64) -> Result<(), LuceneError> {
+    fn fill(&mut self, from_index: i32, to_index: i32, val: i64) -> Result<(), LuceneError> {
         match self {
             MutableEnum::Packed(op) => op.fill(from_index, to_index, val),
             MutableEnum::GrowableW(op) => op.fill(from_index, to_index, val),
